% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird.R
\docType{data}
\name{bird}
\alias{bird}
\title{Data showing counts of various bird species observed in a range of habitats.
#'}
\format{
A tibble with 6 rows and 6 variables:
\describe{
\item{Species}{chr Which bird species}
\item{Garden}{dbl counts of birds seen in garden habitat}
\item{Hedgerow}{dbl counts of birds seen in hedgerow habitat}
\item{Parkland}{dbl counts of birds seen in parkland habitat}
\item{Pasture}{dbl counts of birds seen in pasture habitat}
\item{Woodland}{dbl counts of birds seen in woodland habitat}
}
}
\source{
From Mark Gardener (2017) Statistics for Ecologists Using R and Excel, Pelagic Publishing.
}
\usage{
bird
}
\description{
Data showing counts of various bird species observed in a range of habitats.
#'
}
\keyword{datasets}
