% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callywitn_wind.R
\docType{data}
\name{callywith_wind}
\alias{callywith_wind}
\title{One year of 10 minute average wind speed data}
\format{
A tibble with 52560 rows and 2 variables:
\describe{
\item{\code{t_10min}}{dbl time in units of 10 minutes, starting from the beginning of the year}
\item{\code{v}}{dbl wind speed in m/s, measured as 10 minute averages}
}
}
\source{
Measured at Callywith wind turbine at h = 63 m 2010
}
\usage{
callywith_wind
}
\description{
One year of 10 minute average wind speed data
}
\keyword{datasets}
