% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_turbines.R
\docType{data}
\name{cc_turbines}
\alias{cc_turbines}
\title{Locations of Carland Cross wind turbines in projected CRS 27700}
\format{
A tibble with 10 rows and 2 variables:
\describe{
\item{\code{X}}{dbl Eastings coordinate in metre}
\item{\code{Y}}{dbl Northings coordinate in metre}
}
}
\source{
Carland Cross wind farm, Cornwall
}
\usage{
cc_turbines
}
\description{
Locations of Carland Cross wind turbines in projected CRS 27700
}
\keyword{datasets}
