% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_combined.R
\docType{data}
\name{gm_combined}
\alias{gm_combined}
\title{Gapminder region, population, GDP, and life expectancy data}
\format{
A tibble with 2562 rows and 7 variables:
\describe{
\item{\code{iso_a3}}{chr: ISO alpha-3 country code}
\item{\code{name}}{chr: Country name}
\item{\code{region}}{chr: 4-group Gapminder region}
\item{\code{year}}{dbl: Year}
\item{\code{population}}{dbl: Population}
\item{\code{gdp_per_capita}}{dbl: GDP per capita, PPP dollars}
\item{\code{life_expectancy}}{dbl: Life expectancy in years}
}
}
\source{
Gapminder: https://www.gapminder.org/data/
}
\usage{
gm_combined
}
\description{
Region, population, GDP, and life expectancy data for all countries from 1950 to 2015 every five years.
}
\keyword{datasets}
