% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\docType{data}
\name{growth}
\alias{growth}
\title{Effect of diet and supplement on growth of cows}
\format{
A tibble with 48 rows and 3 variables:
\describe{
\item{\code{supplement}}{chr suppllement: one of agrimore, supergain, control or supersupp}
\item{\code{diet}}{chr diet: one of barley, oats or wheat}
\item{\code{gain}}{dbl gain: weight gain in kg after a fixed period}
}
}
\source{
From Beckerman, Childs and Petchey, 2nd. ed.
}
\usage{
growth
}
\description{
Data of growth in kg of cows fed a combination of one of three diets and one of four supplements.
There are four replicates for each of the twelve combinations of diet and supplement.
}
\keyword{datasets}
