% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladybirds.R
\docType{data}
\name{ladybirds}
\alias{ladybirds}
\title{Tidy data set of counts of red and black ladybirds observed at five rural sites and five urban sites}
\format{
A tibble with 20 rows and 4 variables:
\describe{
\item{\code{habitat}}{chr: rural or urban}
\item{\code{site}}{chr: r1 - r5 and u1 - u5}
\item{\code{morph_colour}}{chr: red or black}
\item{\code{number}}{dbl: count of ladybirds observed}
}
}
\source{
Adapted from Beckerman, Childs and Petchey, 2nd. ed.
}
\usage{
ladybirds
}
\description{
Tidy data set of counts of red and black ladybirds observed at five rural sites and five urban sites
}
\keyword{datasets}
