% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbhR-package.R
\docType{package}
\name{mbhR-package}
\alias{mbhR}
\alias{mbhR-package}
\title{mbhR: Data Package With Data for mbhR classes}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Hunt \email{michael.hunt@cornwall.ac.uk} (\href{https://orcid.org/0009-0005-4394-7178}{ORCID})

}
\keyword{internal}
