% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ozone.R
\docType{data}
\name{ozone}
\alias{ozone}
\title{Data of ozone concentrations in 10 gardens in the east and 10 in the west of a city}
\format{
A tibble with 20 rows and 3 variables:
\describe{
\item{\code{garden.id}}{chr Garden ID G1-G20}
\item{\code{garden.location}}{chr Garden location: East or West}
\item{\code{ozone}}{dbl garden ozone concentration}
}
}
\source{
From Beckerman, Childs and Petchey, 2nd. ed.
}
\usage{
ozone
}
\description{
Data of ozone concentrations in 10 gardens in the east and 10 in the west of a city
}
\keyword{datasets}
