% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant.growth.rate.R
\docType{data}
\name{plants}
\alias{plants}
\title{Tidy data set of masses of male and female squirrels}
\format{
A tibble with 100 rows and 2 variables:
\describe{
\item{\code{soil.moisture.content}}{dbl: soil moisture content}
\item{\code{plant.growth.rate}}{dbl: plant growth rate}
}
}
\source{
From R4All https://github.com/R4All
}
\usage{
plants
}
\description{
Tidy data set of masses of male and female squirrels
}
\keyword{datasets}
