% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squirrels.R
\docType{data}
\name{squirrels}
\alias{squirrels}
\title{Tidy data set of masses of male and female squirrels}
\format{
A tibble with 100 rows and 2 variables:
\describe{
\item{\code{sex}}{chr: FEMALE or MALE}
\item{\code{mass}}{dbl: mass in kg}
}
}
\source{
Adapted from the log transformed data set used in Grafen and Hails (2002), Modern Statistics for the Life Sciences, OUP.
}
\usage{
squirrels
}
\description{
Tidy data set of masses of male and female squirrels
}
\keyword{datasets}
