% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_power_curve.R
\docType{data}
\name{wind_power_curve}
\alias{wind_power_curve}
\title{Power curve for a generic wind turbine}
\format{
A tibble with 25 rows and 2 variables:
\describe{
\item{\code{v}}{dbl wind speed in m/s}
\item{\code{P}}{dbl power output of wind turbine}
}
}
\source{
Typical of many published power curves for large (>1 MW) wind turbines
}
\usage{
wind_power_curve
}
\description{
Power curve for a generic wind turbine
}
\keyword{datasets}
