% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbhR-package.R
\docType{package}
\name{mbhR-package}
\alias{mbhR}
\alias{mbhR-package}
\title{mbhR: Data package with data for mbhR classes}
\description{
This package contains all the data needed for analysis in class R sessions or for self-study.
}
\author{
\strong{Maintainer}: Michael Hunt \email{michael.hunt@cornwall.ac.uk} (\href{https://orcid.org/0009-0005-4394-7178}{ORCID})

}
\keyword{internal}
