% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nashville_carbonates.R
\docType{data}
\name{nashville_carbonates}
\alias{nashville_carbonates}
\title{Geochemical data}
\format{
A tibble with 200 rows and 10 variables:
\describe{
\item{sampleID}{sample ID}
\item{stratPosition}{depth beneath the surface (cm)}
\item{d13C}{C13 condentration}
\item{d18O}{O18 condentration}
\item{Al}{Aluminium concentration}
\item{Ca}{Aluminium concentration}
\item{Fe}{Aluminium concentration}
\item{Mg}{Aluminium concentration}
\item{Mn}{Aluminium concentration}
\item{Si}{Aluminium concentration}
}
}
\source{
{From the GEOL 8370 site at http://stratigrafia.org/8370/data/}

{Theiling, B.P. et al. (2007) ‘Heterogeneity in Geochemical Expression of Subaerial Exposure in Limestones, and Its Implications for Sampling to Detect Exposure Surfaces’, Journal of Sedimentary Research, 77(2), pp. 159–169.} \doi{10.6073/pasta/98b16d7d563f265cb52372c8ca99e60f}
}
\usage{
nashville_carbonates
}
\description{
Geochemical data
}
\keyword{datasets}
